Address thermo-nuclear review items 1–3; track phase-2 work in #23–#25.
Extract shared partial-content emitter and canonical secret scrubbing, merge transcript activity/standard specs into one registry, and dedupe record/text helpers.
