+++ src/cursor-record-utils.ts
@@ -2,6 +2,14 @@ export function asRecord(value: unknown): Record<string, unknown> | undefined {
 	return value && typeof value === "object" && !Array.isArray(value) ? (value as Record<string, unknown>) : undefined;
 }
 
+export function getField(value: unknown, field: string): unknown {
+	return asRecord(value)?.[field];
+}
+
+export function hasUsableText(value: string | undefined): value is string {
+	return typeof value === "string" && value.trim().length > 0;
+}
+
 export function getFirstStringByKeys(
 	record: Record<string, unknown> | undefined,
 	keys: readonly string[],
