+++ src/cursor-transcript-tool-specs.ts
@@ -1,7 +1,8 @@
-import { CURSOR_REPLAY_ACTIVITY_TOOL_NAME, getCursorReplayDisplayLabel } from "./cursor-tool-names.js";
+import { CURSOR_REPLAY_ACTIVITY_TOOL_NAME, getCursorReplayDisplayLabel, type CursorReplayLegacyToolName } from "./cursor-tool-names.js";
 import { resolveCursorEditDiff } from "./cursor-edit-diff.js";
 import {
 	asRecord,
+	firstNonEmptyLine,
 	formatDisplayPath,
 	formatDiffString,
 	formatError,
@@ -9,16 +10,12 @@ import {
 	getString,
 	limitText,
 	stringifyUnknown,
+	truncateArg,
 	type CursorPiToolDisplay,
 	type NormalizedResult,
+	type PiToolDisplayResult,
 	type TranscriptOptions,
 } from "./cursor-transcript-utils.js";
-import {
-	buildCursorActivityDisplayArgs,
-	buildReplaySummaryDisplay,
-	buildStandardActivityReplayDisplay,
-	textToolResult,
-} from "./cursor-transcript-activity-display.js";
 import {
 	buildCursorEditActivityDisplayArgs,
 	buildFindDisplayArgs,
@@ -27,6 +24,7 @@ import {
 	buildReadDisplayArgs,
 	buildShellDisplayArgs,
 	buildWriteDisplayArgs,
+	collectTaskText,
 	formatDelete,
 	formatEdit,
 	formatFallback,
@@ -44,10 +42,21 @@ import {
 	formatWrite,
 	formatNativeReadDisplayContent,
 	getCursorWriteArgContent,
+	getGenerateImageDisplayPath,
+	getGenerateImagePath,
 	getGlobBody,
 	getGrepBody,
 	getLsBody,
+	getReadLintDiagnostics,
+	getReadLintPaths,
 	getShellOutput,
+	getTaskDescription,
+	getTodoItems,
+	getTodoTotalCount,
+	inferImageMimeType,
+	summarizePlan,
+	summarizeTask,
+	summarizeTodos,
 } from "./cursor-transcript-tool-formatters.js";
 
 export interface ToolDisplayContext {
@@ -58,9 +67,81 @@ export interface ToolDisplayContext {
 	options: TranscriptOptions;
 }
 
+interface ActivityReplaySpec {
+	labelKey: CursorReplayLegacyToolName;
+	buildActivityArgs: (context: ToolDisplayContext) => Record<string, unknown>;
+	buildActivitySummary: (context: ToolDisplayContext) => string | undefined;
+	buildDetails: (context: ToolDisplayContext, contentText: string) => Record<string, unknown>;
+}
+
 interface ToolDisplaySpec {
 	formatTranscript: (context: ToolDisplayContext) => string;
 	buildPiToolDisplay: (context: ToolDisplayContext) => CursorPiToolDisplay;
+	activityReplay?: ActivityReplaySpec;
+}
+
+function textToolResult(text: string, details?: unknown): PiToolDisplayResult {
+	return { content: [{ type: "text", text }], details };
+}
+
+function buildCursorActivityDisplayArgs(
+	args: Record<string, unknown>,
+	activityTitle: string,
+	activitySummary: string | undefined,
+): Record<string, unknown> {
+	const trimmedSummary = activitySummary?.trim();
+	return {
+		...args,
+		activityTitle,
+		...(trimmedSummary ? { activitySummary: trimmedSummary } : {}),
+	};
+}
+
+function buildReplaySummaryDisplay(
+	toolName: string,
+	args: Record<string, unknown>,
+	result: NormalizedResult,
+	contentText: string,
+	details: Record<string, unknown>,
+): CursorPiToolDisplay {
+	const isError = result.status === "error";
+	const summary = isError ? formatError(result.error) : firstNonEmptyLine(contentText);
+	return {
+		toolName,
+		args,
+		result: textToolResult(contentText, {
+			...details,
+			summary: details.summary ?? summary,
+			expandedText: details.expandedText ?? contentText,
+		}),
+		isError,
+	};
+}
+
+function buildActivityReplayDisplay(cursorToolName: string, spec: ToolDisplaySpec, context: ToolDisplayContext): CursorPiToolDisplay {
+	const activity = spec.activityReplay;
+	if (!activity) throw new Error(`Missing activity replay spec for ${cursorToolName}`);
+	const activityTitle = getCursorReplayDisplayLabel(activity.labelKey);
+	const activitySummary = activity.buildActivitySummary(context);
+	const activityArgs = buildCursorActivityDisplayArgs(
+		activity.buildActivityArgs(context),
+		activityTitle,
+		activitySummary,
+	);
+	const contentText = spec.formatTranscript(context).trimEnd();
+	const details = activity.buildDetails(context, contentText);
+	return buildReplaySummaryDisplay(
+		CURSOR_REPLAY_ACTIVITY_TOOL_NAME,
+		activityArgs,
+		context.result,
+		contentText,
+		{
+			cursorToolName,
+			title: activityTitle,
+			summary: context.result.status === "error" ? undefined : details.summary ?? activitySummary,
+			...details,
+		},
+	);
 }
 
 function buildGenericPiToolDisplay(context: ToolDisplayContext): CursorPiToolDisplay {
@@ -221,31 +302,129 @@ const TOOL_DISPLAY_SPECS: Record<string, ToolDisplaySpec> = {
 	},
 	delete: {
 		formatTranscript: ({ args, result, options }) => formatDelete(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("delete", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("delete", TOOL_DISPLAY_SPECS.delete, context),
+		activityReplay: {
+			labelKey: "cursor_delete",
+			buildActivityArgs: ({ args, options }) => {
+				const displayPath = typeof args.path === "string" ? formatDisplayPath(args.path, options.cwd) : undefined;
+				return displayPath ? { path: displayPath } : {};
+			},
+			buildActivitySummary: ({ args, options }) => {
+				const displayPath = typeof args.path === "string" ? formatDisplayPath(args.path, options.cwd) : undefined;
+				return displayPath ?? "file";
+			},
+			buildDetails: ({ args, result, options }) => {
+				const displayPath = typeof args.path === "string" ? formatDisplayPath(args.path, options.cwd) : undefined;
+				const value = asRecord(result.value);
+				return {
+					path: displayPath,
+					fileSize: getNumber(value, "fileSize"),
+					summary: result.status === "error" ? undefined : displayPath ? `deleted ${displayPath}` : "deleted file",
+				};
+			},
+		},
 	},
 	readLints: {
 		formatTranscript: ({ args, result, options }) => formatReadLints(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("readLints", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("readLints", TOOL_DISPLAY_SPECS.readLints, context),
+		activityReplay: {
+			labelKey: "cursor_read_lints",
+			buildActivityArgs: ({ args, result, options }) => {
+				const paths = getReadLintPaths(args, result, options);
+				const diagnosticCount = getReadLintDiagnostics(result, options).length;
+				return { paths, diagnosticCount };
+			},
+			buildActivitySummary: ({ args, result, options }) => {
+				const paths = getReadLintPaths(args, result, options);
+				const diagnosticCount = getReadLintDiagnostics(result, options).length;
+				return `${diagnosticCount} diagnostic${diagnosticCount === 1 ? "" : "s"}${paths.length > 0 ? ` in ${paths.join(", ")}` : ""}`;
+			},
+			buildDetails: () => ({}),
+		},
 	},
 	updateTodos: {
 		formatTranscript: ({ args, result, options }) => formatTodos(args, result, options, "updateTodos"),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("updateTodos", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("updateTodos", TOOL_DISPLAY_SPECS.updateTodos, context),
+		activityReplay: {
+			labelKey: "cursor_update_todos",
+			buildActivityArgs: ({ args, result }) => {
+				const todos = getTodoItems(args, result);
+				return { totalCount: getTodoTotalCount(args, result, todos) };
+			},
+			buildActivitySummary: ({ args, result }) => summarizeTodos(args, result),
+			buildDetails: () => ({}),
+		},
 	},
 	createPlan: {
 		formatTranscript: ({ args, result, options }) => formatPlan(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("createPlan", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("createPlan", TOOL_DISPLAY_SPECS.createPlan, context),
+		activityReplay: {
+			labelKey: "cursor_create_plan",
+			buildActivityArgs: ({ args, result }) => {
+				const todos = getTodoItems(args, result);
+				return { totalCount: getTodoTotalCount(args, result, todos) };
+			},
+			buildActivitySummary: ({ args, result }) => summarizePlan(args, result),
+			buildDetails: () => ({}),
+		},
 	},
 	task: {
 		formatTranscript: ({ args, result, options }) => formatTask(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("task", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("task", TOOL_DISPLAY_SPECS.task, context),
+		activityReplay: {
+			labelKey: "cursor_task",
+			buildActivityArgs: ({ args, result }) => {
+				const description = getTaskDescription(args, result);
+				return { description: truncateArg(description) };
+			},
+			buildActivitySummary: ({ args, result }) => {
+				const description = getTaskDescription(args, result);
+				return summarizeTask(description, collectTaskText(result));
+			},
+			buildDetails: () => ({}),
+		},
 	},
 	generateImage: {
 		formatTranscript: ({ args, result, options }) => formatGenerateImage(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("generateImage", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("generateImage", TOOL_DISPLAY_SPECS.generateImage, context),
+		activityReplay: {
+			labelKey: "cursor_generate_image",
+			buildActivityArgs: ({ args }) => {
+				const prompt = getString(args, "prompt") ?? getString(args, "description") ?? "image";
+				return { prompt: truncateArg(prompt) };
+			},
+			buildActivitySummary: ({ args, result, options }) => {
+				const prompt = getString(args, "prompt") ?? getString(args, "description") ?? "image";
+				const imageDisplayPath = getGenerateImageDisplayPath(args, result, options);
+				return imageDisplayPath ?? truncateArg(prompt);
+			},
+			buildDetails: ({ args, result, options }, contentText) => {
+				const imagePath = getGenerateImagePath(args, result);
+				const imageDisplayPath = getGenerateImageDisplayPath(args, result, options);
+				return {
+					imagePath,
+					imageDisplayPath,
+					imageMimeType: inferImageMimeType(imagePath),
+					summary: result.status === "error" ? undefined : imageDisplayPath ? `saved ${imageDisplayPath}` : "image generated",
+					expandedText: contentText,
+				};
+			},
+		},
 	},
 	mcp: {
 		formatTranscript: ({ args, result, options }) => formatMcp(args, result, options),
-		buildPiToolDisplay: (context) => buildStandardActivityReplayDisplay("mcp", context) ?? buildGenericPiToolDisplay(context),
+		buildPiToolDisplay: (context) => buildActivityReplayDisplay("mcp", TOOL_DISPLAY_SPECS.mcp, context),
+		activityReplay: {
+			labelKey: "cursor_mcp",
+			buildActivityArgs: ({ args }) => {
+				const toolName = getString(args, "toolName") ?? "mcp";
+				return { toolName: truncateArg(toolName) };
+			},
+			buildActivitySummary: ({ args }) => truncateArg(getString(args, "toolName") ?? "mcp"),
+			buildDetails: ({ result }, contentText) => ({
+				summary: result.status === "error" ? undefined : firstNonEmptyLine(contentText) ?? "MCP result captured",
+			}),
+		},
 	},
 };
